requirements.txt
faker
pyyaml
requests